requirements.txt
fastapi
uvicorn 
python-slugify
vaderSentiment